Support for rethinkdbdash driver
Fixes #9
