+++ package.json
@@ -20,8 +20,8 @@
   },
   "devDependencies": {
     "mocha": "^2.2.4",
-    "protolog": "0.0.3",
     "rethinkdb": "^2.0.0",
+    "rethinkdbdash": "^2.2.17",
     "should": "^6.0.1"
   },
   "repository": {
